Convenient access to the CCLE proteome data

Functionality for storing the protein expression data from the Cancer Cell Line Encyclopedia (CCLE) in a SummarizedExperiment.

ccleProteome2SummarizedExperiment(df, cell.line = "HCT116")

Arguments

df

a data.frame storing the CCLE protein expression data with one measurement in each row. Typically obtained from ExperimentHub. See examples.

cell.line

character. One or more cell line IDs such as "HCT116" (human colon cancer cell line 116). Use NULL to not subset by cell line. Defaults to "HCT116", which will then subset the df to measurements for HCT116 only.

Value

A SummarizedExperiment storing protein expression data for the specified cell line(s).

References

CCLE proteomics: https://gygi.hms.harvard.edu/publications/ccle.html

Examples

  # Connect to ExperimentHub
  eh <- ExperimentHub::ExperimentHub()
#> snapshotDate(): 2022-10-24

  # Obtain CCLE proteome data frame
  AnnotationHub::query(eh, c("gygi", "depmap"))
#> ExperimentHub with 1 record
#> # snapshotDate(): 2022-10-24
#> # names(): EH3459
#> # package(): depmap
#> # $dataprovider: Broad Institute
#> # $species: Homo sapiens
#> # $rdataclass: tibble
#> # $rdatadateadded: 2020-05-19
#> # $title: proteomic_20Q2
#> # $description: Quantitative profiling of 12399 proteins in 375 cell lines, ...
#> # $taxonomyid: 9606
#> # $genome: 
#> # $sourcetype: CSV
#> # $sourceurl: https://gygi.med.harvard.edu/sites/gygi.med.harvard.edu/files/...
#> # $sourcesize: NA
#> # $tags: c("ExperimentHub", "ExperimentData", "ReproducibleResearch",
#> #   "RepositoryData", "AssayDomainData", "CopyNumberVariationData",
#> #   "DiseaseModel", "CancerData", "BreastCancerData", "ColonCancerData",
#> #   "KidneyCancerData", "LeukemiaCancerData", "LungCancerData",
#> #   "OvarianCancerData", "ProstateCancerData", "OrganismData",
#> #   "Homo_sapiens_Data", "PackageTypeData", "SpecimenSource",
#> #   "CellCulture", "Genome", "Proteome", "StemCell", "Tissue") 
#> # retrieve record with 'object[["EH3459"]]' 
  ccle.prot <- eh[["EH3459"]]
#> snapshotDate(): 2022-10-24
#> see ?depmap and browseVignettes('depmap') for documentation
#> loading from cache
  ccle.prot <- as.data.frame(ccle.prot)

  # Turn into a SummarizedExperiment
  se <- ccleProteome2SummarizedExperiment(ccle.prot)