Identify CORUM complexes that have a subunit of interest

Screens a list of graph instances storing CORUM protein complex data for a subunit of choice.

hasSubunit(glist, subunit, id.type = "SYMBOL")

Arguments

glist
A list of graphs storing CORUM complexes. Typically obtained via corum2graphlist.

subunit
character. A gene ID corresponding to the subunit of interest.

id.type
character. Gene ID type of the given subunit. Defaults to "SYMBOL".

Value

A logical vector indicating which graphs have a node with the given subunit.

Examples

# (1) Obtain the core set of CORUM complexes ...
core <- getCorum(set = "core")

#> Using cached version from 2021-06-25 05:24:35

# (2) ... turn into a list of graphs ...
core.glist <- corum2graphlist(core)

# (3) .. check for a particular subunit of interest
has.cdk2 <- hasSubunit(core.glist, subunit = "CDK2")