Map experimental data onto a graph

Functionality for mapping experimental data stored in a SummarizedExperiment onto a graph object.

mapSummarizedExperimentOntoGraph(
  gr,
  se,
  col.names = NULL,
  rowdata.cols = NULL,
  prefix = ""
)

Arguments

gr

an object of class graph.

se

an object of class SummarizedExperiment.

col.names

character. Column names of se for which assay data should be mapped onto the nodes of gr. Defaults to NULL which will then use all column names of se.

rowdata.cols

character. Column names of rowData(se) which should be mapped onto the nodes of gr. Defaults to NULL which will then use all column names of rowData(se).

prefix

character. Informative prefix that should be pasted together with the selected col.names and rowdata.cols to allow easy identification of columns of interest when mapping from multiple experimental datasets.

Value

An object of class graph.

Examples

# (1) Obtain the latest version of the 293T PPI network ...
bp.293t <- getBioPlex(cell.line = "293T", version = "3.0")
#> Using cached version from 2022-11-13 22:56:05

# (2) ... and turn into a graph
bp.gr <- bioplex2graph(bp.293t)

# (3) Obtain the BioPlex3 proteome data ...
se <- getBioplexProteome()
#> Using cached version from 2022-11-13 22:56:43

# (4) ... and map onto the graph
bp.gr <- mapSummarizedExperimentOntoGraph(bp.gr, se)